added native linking to phone number and email in ConnectionView
