+++ src/components/CaseListComponent.js
@@ -31,19 +31,7 @@ import {
     Badge,
     SocialIcon
 } from "react-native-elements";
-
-const formatTelephone = (telephone) => {
-    if (telephone !== null) {
-        let removed = telephone.slice(2)
-        let areaCode = removed.slice(0, 3)
-        let prefix = removed.slice(3, 6)
-        let lineNumber = removed.slice(6, 10)
-        return `(${areaCode}) ${prefix}-${lineNumber}`
-    } else {
-        return ''
-    }
-    
-}
+import formatTelephone from '../helpers/formatTelephone.js'
 
 const CaseListComponent = (props) => {
     return (
